Correction of the repository. Creation of the register request in UsersController.

+++ src/main/java/com/epsi/marche_malin_api/entities/Chat.java
@@ -17,7 +17,7 @@ import lombok.Setter;
 public class Chat {
     @Id
     @Column(name = "id_chat")
-    private int idChat;
+    private Long idChat;
 
     @ManyToOne
     @JoinColumn(name = "user1")
